Move vscode-mdx extension into a workspace
Everything that is specific to the VSCode extension, is now part of the
`vscode-mdx` workspace. This includes most things from before
IntelliSense was added, but also the VSCode language client code.

Because most of this is licensed by @JounQin, I copied the license from
the project root into the workspace.

The project root readme was also moved into the `vscode-mdx` workspace.
A new one was written for the project root.