Support remark plugins in the language server (#272)
Syntax plugins can be added by adding an `mdx` section in your
`tsconfig.json` file. This is then parsed following the same format as
`unified-engine` [configuration][].

For example, to support [frontmatter][] and [GFM][], add the following
section to `tsconfig.json`:

```jsonc
{
  "compilerOptions": {
    // …
  },
  "mdx": {
    "plugins": [
      "remark-frontmatter",
      "remark-gfm"
    ]
  }
}
```

This also adds the JSON schema for the new field in `tsconfig.json`.