Remove backticks from `autoClosingPairs` and `surroundingPairs` to prevent doubling them up
Doubling up backticks happens when trying to write a code block in MDX.

This breaks backtick auto-pairing even when writing inline code span, but arguably, it is much less annoying to have to add the matching backtick manually in the case of an inline code span than it is to have to remove it twice every time a code block is being written (in the opening and closing triple-backtick lines).

If you know how to give this file more context so it could only be disabled for code blocks, let me know. Althoug I am not sure it is super desirable to have backtick completions even for inline code spans. But that's just my preference. The original issue reported is facing issues with auto-pairing in inline code spans so I think it is worth just removing this for backticks altogether.

I have added a bit of documentation on how to run the extension. I've written a few VS Code extensions myself but it has been a long time and I've forgotten everything so I had to figure out the instructions. Seems like a waste to not contribute them if there's interest in taking them in.

Initially I thought the way to fix this would be to spy on the LSP messages and find the one responsible for handling the backtick and completing the matching pair, hence why the instructions contain information about LSP messages. In the end it was a much more low level feature so I did not find anything useful in the LSP messages.

Resolves #231