Fix various bugs related to shadow positions (#300)
MDX content is converted to JavaScript. TypeScript then uses this
JavaScript handle requests. It then returns positional information that
needs to be mapped back to the original MDX content.

This was previously not handled properly, leading to mostly false
positive results. This is now handled better by stripping results that
reference locations that can’t be mapped back to the original MDX
content.