Handle dragging uri-list into MDX files (#346)
A mime type of `uri-list` is for example files dragged from the VSCode
tree view.

Images are turnes into Markdown images. These are detected based on a
hardcoded list of file extensions. Other URIs are inserted as-is.

If the dragged URI uses the same scheme as the document it was dragged
in, a relative path is inserted. Otherwise the full URI is used.

Closes #322