Enhance how the Props type is used (#366)
If the Props type isn’t defined, it now falls back to an empty shape
instead of any. Also all keys are marked as readonly. As a result, props
are no longer represented using the string `Props`, but an inline
representation of the shape of the props.

Given that the upcoming support for an MDX layout also requires a
patched version of `Props`, this new setup is more consistent.

Also the `props` description now contains a link to the props page on
the MDX website.