Add minimal support for the components prop
This defines the `_components` variable in `_createMdxContent`. This
variable contains all components injected through the `components` prop,
a reference to `props`, and all local variables. All MDX JSX tags are
prefixed with `_components.` in the virtual code.

As a result, components declared in the `components` prop are allowed.
Components from both the `components` prop and local variables are
displayed in the autocomplete.

A downside of this approach is that documentation is lost for local
components. To mitigate this, only unknown JSX tags are prefixed.

Only MDX JSX tags are handled yet. JSX from estree not yet.

Refs #260