Move basic usage instructions to the repo root
The instructions on how to configure TypeScript or remark plugins is not
specific to the language server. It also applies to the VS Code plugin
and the upcoming TypeScript plugin, and even a bit to the language service.

These instructions are moved to the reposity root readme, so the user is
more likely to encounter them faster.