Add support for some popular transformer plugins
This adds support for the following transformer plugins:
- `recma-export-filepath`
- `rehype-mdx-title`
- `remark-mdx-frontmatter`

These plugins all follow the same pattern; they expose some form of
metatada as an ESM export. I think these are some of the most popular
plugins that actually affect the editor experience.

I don’t like that support for these plugins is hardcoded. This is only
due to the lack of an alternative. For now I welcome hardcoded support
for other plugins.

I intend to iterate on this to provide a better editor experience for
YAML frontmatter.

Refs #297