Drop support for Node.js 18 (#499)
The asynchronous TypeScript plugin keeps causing issues. The goal is to
move away from that and make it synchronous. Node.js 20.19 adds support
for `require(esm)`, which is a great first step for our use case.

VSCode currently ships with Node.js 20.18. However, this is still fine,
as dependencies are bundled in the VSCode extension.