Resolve remark plugins synchronously
TypeScript plugins only support a synchronous API. Volar has a way to
support asynchronous plugins, but it’s hacky and keeps breaking. This
change makes the loading of remark plugins synchronous, allowing us to
make the TypeScript plugin synchronous as well.

For now jiti is used to load plugins synchronously. As soon as VSCode
ships with a Node.js version that supports `require(esm)`, we can
replace jiti with `createRequire()`.